backend/requirements.txt
attrs==21.2.0
asgiref==3.2.10
certifi==2021.5.30
charset-normalizer==2.0.4
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-colorfield==0.6.3
drf-extra-fields==3.2.1
djoser==2.1.0
idna==3.2
iniconfig==1.1.1
numpy==1.21.1
packaging==21.0
pandas==1.3.1
pluggy==0.13.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary==2.8.6